+++ doc-app/doc-entry/types.d.ts
@@ -1,8 +1,7 @@
 declare module '*.md' {
   import {LoaderRecivedData} from '@wfh/doc-ui-common/isom/md-types';
-  const html: string;
-  const toc: LoaderRecivedData['toc'];
-  export {html, toc};
+  const data: LoaderRecivedData;
+  export default data;
   global {
     interface ImportMeta {
       webpackHot: any;