+++ doc-app/doc-ui-common/ts/markdown-processor.ts
@@ -1,11 +1,11 @@
 import * as rx from 'rxjs';
 import {log4File} from '@wfh/plink';
 import {createWorkerControl, fork, WorkerControl} from '@wfh/reactivizer/dist/fork-join/node-worker';
-import {ActionMeta, serializeAction, str2ArrayBuffer, arrayBuffer2str} from '@wfh/reactivizer';
+import {timeoutLog, ActionMeta, str2ArrayBuffer, arrayBuffer2str} from '@wfh/reactivizer';
 import md5 from 'md5';
 import MarkdownIt from 'markdown-it';
 import highlight from 'highlight.js';
-import type {DefaultTreeAdapterMap} from 'parse5';
+import {parse as parseHtml, DefaultTreeAdapterMap} from 'parse5';
 import {TOC} from '../isom/md-types';
 import {ChildNode, Element, lookupTextNodeIn, createTocTree} from './markdown-processor-helper';
 
@@ -17,6 +17,7 @@ type MdInputActions = {
   processFileDone(res: {resultHtml: ArrayBuffer; toc: TOC[]; mermaid: ArrayBuffer[]; transferList: ArrayBuffer[]}): void;
   /** Consumer should dispatach to be related to "resolveImage" event */
   imageResolved(resultUrl: string): void;
+  linkResolved(resultUrl: string): void;
   /** Consumer should dispatch */
   anchorLinkResolved(url: string): void;
 };
@@ -25,6 +26,8 @@ export type MdOutputEvents = {
   processFileDone: MdInputActions['processFileDone'];
   /** Consumer program should react on this event */
   imageToBeResolved(imgSrc: string, mdFilePath: string): void;
+  /** Consumer program should react on this event */
+  linkToBeResolved(urlSrc: string, mdFilePath: string): void;
   /** Consumer should react and dispatach "anchorLinkResolved" */
   anchorLinkToBeResolved(linkSrc: string, mdFilePath: string): void;
 
@@ -49,13 +52,15 @@ const md = new MarkdownIt({
 export type MarkdownProcessor = WorkerControl<MdInputActions, MdOutputEvents>;
 
 export const markdownProcessor: MarkdownProcessor = createWorkerControl<MdInputActions, MdOutputEvents>({
-  name: 'markdownProcessor', debug: true,
+  name: 'markdownProcessor',
+  debug: true,
+  debugExcludeTypes: ['wait', 'stopWaiting'],
   log(...msg) {
     log.info(...msg);
   }
 });
 
-const {r, i, o, outputTable} = markdownProcessor;
+const {r, i, o} = markdownProcessor;
 
 r('forkProcessFile -> fork processFile, processFileDone', i.pt.forkProcessFile.pipe(
   rx.mergeMap(async ([m, content, file]) => {
@@ -68,10 +73,10 @@ r('forkProcessFile -> fork processFile, processFileDone', i.pt.forkProcessFile.p
 ));
 
 r('processFile -> processFileDone', i.pt.processFile.pipe(
-  rx.combineLatestWith(import('parse5')),
-  rx.mergeMap(([[m, content, file], parse5]) => {
+  rx.tap(() => {o.dp.log('react to processFile'); }),
+  rx.mergeMap(([m, content, file]) => {
     const html = md.render(arrayBuffer2str(content));
-    const doc = parse5.parse(html, {sourceCodeLocationInfo: true});
+    const doc = parseHtml(html, {sourceCodeLocationInfo: true});
     const content$ = dfsAccessElement(m, html, file, doc);
     return content$.pipe(
       rx.map(([content, toc, mermaidCodes]) => {
@@ -83,23 +88,19 @@ r('processFile -> processFileDone', i.pt.processFile.pipe(
   }),
   rx.catchError(err => {
     log.error(err);
+    o.dp.log(err);
     // o.dpf.processFileDone({toc: [], content: (err as Error).toString()});
     return rx.EMPTY;
   })
 ));
 
-r('imageToBeResolved -> main thread port postMessage', o.at.imageToBeResolved.pipe(
-  rx.withLatestFrom(outputTable.l.workerInited),
-  rx.tap(([resolveAction, [, , , port]]) => {
-    if (port) {
-      o.dp.log('pass imageToBeResolved action to main thread');
-      port.postMessage(serializeAction(resolveAction));
-    }
-  })
+i.dp.setLiftUpActions(rx.merge(
+  o.at.imageToBeResolved,
+  o.at.linkToBeResolved
 ));
 
 function dfsAccessElement(
-  processFileActionMeta: ActionMeta,
+  _processFileActionMeta: ActionMeta,
   sourceHtml: string,
   file: string,
   root: DefaultTreeAdapterMap['document']
@@ -168,6 +169,23 @@ function dfsAccessElement(
           text: lookupTextNodeIn(el),
           id: hash
         });
+      } else if (nodeName === 'a') {
+        const hrefAttr = el.attrs.find(attr => attr.name === 'href');
+        if (hrefAttr?.value && hrefAttr.value.startsWith('.')) {
+          output.push(sourceHtml.slice(htmlOffset, el.sourceCodeLocation!.attrs!.href!.startOffset + 'href="'.length));
+          htmlOffset = el.sourceCodeLocation!.attrs!.href!.endOffset - 1;
+          return output.push(rx.merge(
+            new rx.Observable<never>(() => {
+              o.dp.wait();
+            }),
+            o.do.linkToBeResolved(i.at.linkResolved, hrefAttr?.value, file)
+          ).pipe(
+            rx.take(1),
+            timeoutLog(3000, () => log.warn('link resolve timeout')),
+            rx.map(([, url]) => url),
+            rx.finalize(() => o.dp.stopWaiting())
+          ));
+        }
       } else if (el.childNodes) {
         chr.next(el.childNodes);
       }