+++ main/packages/reactivizer/es/fork-join/node-worker.d.ts
@@ -11,7 +11,7 @@ import { ReactorCompositeOpt } from '../epic';
 import { ForkWorkerInput, ForkWorkerOutput, WorkerControl } from './types';
 export { fork } from './common';
 export { WorkerControl } from './types';
-export declare function createWorkerControl<I extends ActionFunctions = Record<string, never>, O extends ActionFunctions = Record<string, never>, LI extends ReadonlyArray<keyof I> = readonly [], LO extends ReadonlyArray<keyof O> = readonly []>(opts?: ReactorCompositeOpt<ForkWorkerInput & ForkWorkerOutput & I & O>): WorkerControl<I, O, LI, LO>;
+export declare function createWorkerControl<I extends ActionFunctions = Record<string, never>, O extends ActionFunctions = Record<string, never>, LI extends ReadonlyArray<keyof I> = readonly [], LO extends ReadonlyArray<keyof O> = readonly []>(opts?: ReactorCompositeOpt<ForkWorkerInput & ForkWorkerOutput & I, ForkWorkerOutput & O>): WorkerControl<I, O, LI, LO>;
 export type ForkTransferablePayload<T = unknown> = {
     content: T;
     transferList: (ArrayBuffer | MessagePort | fsPromises.FileHandle | X509Certificate | Blob)[];