+++ main/packages/reactivizer/es/stream-core.js
@@ -9,7 +9,7 @@ export class ControllerCore {
         this.actionUpstream = new rx.Subject();
         this.interceptor$ = new rx.BehaviorSubject(a => a);
         this.typePrefix = '#' + SEQ++ + ' ';
-        this.logPrefix = this.typePrefix;
+        this.logPrefix = ''; // TODO: a better identity to distinguish threads
         this.dispatcher = {};
         this.dispatcherFor = {};
         this.actionSubDispatcher = new rx.Subject();
@@ -69,7 +69,7 @@ export class ControllerCore {
     }
     /** change the "name" as previous specified in CoreOptions of constructor */
     setName(name) {
-        this.logPrefix = name ? `[${this.typePrefix}${name}] ` : this.typePrefix;
+        this.logPrefix = name !== null && name !== void 0 ? name : this.typePrefix;
     }
     dispatchFactory(type) {
         if (has.call(this.dispatcher, type)) {